First working version of parser
The idea is to control the flow of the pipeline by a YAML specification
file. The pipeline is made out of multiple individual routines, where
the output of one is the input of the next. This way the user can
control which routine goes first, etc.

First working version of specification verification and parser.