One process On set of bands
I removed the possibility of having multiple bodies, for the reason that
it will be too much to parallelize per body and then per set of bands.
Furthermore I think there is no need to run the exact pipeline on
multiple bodies (within IMEGIN). If needed, one can run multiple
instances of the pipeline with different yaml files.