OOP implem for class creation and better file_mng
The pipeline is now written in an OOP manner that makes it easier to
read and maintain the library. Now, the creation of Pipeline() will
either create a PipelineParallel() or PipelineSequential() depending on
the parallel key in the yaml specification file.

Made the file_manager better by getting around some edge cases of
parsing.

To make mypy pass, had to correct some type hints and add the
__init__.py files in util/ and pipeline/